♻️ refactor: break the revision-write import cycle
fallow flagged 3 circular dependencies: issue-write ↔ revision-write and
case-study-write ↔ revision-write (the latter also via issue-write). The cycle
existed because revision-write owned both the snapshot/record primitives (which
the writers call to log born-approved revisions) and the apply lifecycle (which
calls updateIssue/updateCaseStudy back in the writers).

Extract the dependency-free primitives — editable*Snapshot, diffSnapshots,
staleConflictFields, recordRevision, serializeRevision, canViewRevision and
their types — into a new leaf module server/utils/revision-record.ts.
issue-write/case-study-write now import those straight from the leaf, so they no
longer depend on revision-write. revision-write keeps the propose/apply/decide/
withdraw lifecycle and re-exports the leaf (`export * from './revision-record'`)
so all other importers (API routes, MCP tools, tests) are unchanged.

Verified: fallow reports 0 cycles, vue-tsc, 68 unit tests, vize fmt --check, and
the Worker build all pass.
